# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Safe Place API",
    "version": "1.0.0"
  },
  "schemes": [
    "http"
  ],
  "host": "localhost:3000",
  "basePath": "/api",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/comments": {
      "get": {
        "summary": "Get all comments",
        "operationId": "getAllComments",
        "tags": ["Comments"],
        "responses": {
          "200": {
            "description": "A list of comments",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Comment"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new comment",
        "operationId": "createComment",
        "tags": ["Comments"],
        "parameters": [
          {
            "in": "body",
            "name": "comment",
            "description": "Comment object to create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewComment"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Comment created successfully",
            "schema": {
              "$ref": "#/definitions/Comment"
            }
          }
        }
      }
    },
    "/comments/{id}": {
      "get": {
        "summary": "Get comment by ID",
        "operationId": "getCommentById",
        "tags": ["Comments"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Comment details",
            "schema": {
              "$ref": "#/definitions/Comment"
            }
          },
          "404": {
            "description": "Comment not found"
          }
        }
      },
      "put": {
        "summary": "Update comment by ID",
        "operationId": "updateComment",
        "tags": ["Comments"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "comment",
            "description": "Updated comment object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewComment"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comment updated successfully",
            "schema": {
              "$ref": "#/definitions/Comment"
            }
          },
          "404": {
            "description": "Comment not found"
          }
        }
      },
      "delete": {
        "summary": "Delete comment by ID",
        "operationId": "deleteComment",
        "tags": ["Comments"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Comment deleted successfully"
          },
          "404": {
            "description": "Comment not found"
          }
        }
      }
    },
...
    "/api_results": {
      "get": {
        "summary": "Get all API results",
        "operationId": "getAllAPIResults",
        "tags": ["API_Results"],
        "responses": {
          "200": {
            "description": "A list of API results",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/APIResult"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new API result",
        "operationId": "createAPIResult",
        "tags": ["API_Results"],
        "parameters": [
          {
            "in": "body",
            "name": "api_result",
            "description": "API result object to create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewAPIResult"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "API result created successfully",
            "schema": {
              "$ref": "#/definitions/APIResult"
            }
          }
        }
      }
    },
    "/api_results/{id}": {
      "get": {
        "summary": "Get API result by ID",
        "operationId": "getAPIResultById",
        "tags": ["API_Results"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "API result details",
            "schema": {
              "$ref": "#/definitions/APIResult"
            }
          },
          "404": {
            "description": "API result not found"
          }
        }
      },
      "put": {
        "summary": "Update API result by ID",
        "operationId": "updateAPIResult",
        "tags": ["API_Results"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "api_result",
            "description": "Updated API result object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewAPIResult"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API result updated successfully",
            "schema": {
              "$ref": "#/definitions/APIResult"
            }
          },
          "404": {
            "description": "API result not found"
          }
        }
      },
      "delete": {
        "summary": "Delete API result by ID",
        "operationId": "deleteAPIResult",
        "tags": ["API_Results"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "API result deleted successfully"
          },
          "404": {
            "description": "API result not found"
          }
        }
      }
    },
    "/categories": {
      "get": {
        "summary": "Get all categories",
        "operationId": "getAllCategories",
        "tags": ["Categories"],
        "responses": {
          "200": {
            "description": "A list of categories",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Category"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new category",
        "operationId": "createCategory",
        "tags": ["Categories"],
        "parameters": [
          {
            "in": "body",
            "name": "category",
            "description": "Category object to create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewCategory"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Category created successfully",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "summary": "Get category by ID",
        "operationId": "getCategoryById",
        "tags": ["Categories"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Category details",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          },
          "404": {
            "description": "Category not found"
          }
        }
      },
      "put": {
        "summary": "Update category by ID",
        "operationId": "updateCategory",
        "tags": ["Categories"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "category",
            "description": "Updated category object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewCategory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category updated successfully",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          },
          "404": {
            "description": "Category not found"
          }
        }
      },
      "delete": {
        "summary": "Delete category by ID",
        "operationId": "deleteCategory",
        "tags": ["Categories"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Category deleted successfully"
          },
          "404": {
            "description": "Category not found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Post": {
      "type": "object",
      "properties": {
        "Id_Post": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "visibility": {
          "type": "string"
        },
        "toxic_score": {
          "type": "number",
          "format": "decimal"
        },
        "image_exists": {
          "type": "boolean"
        },
        "image_b64": {
          "type": "string",
          "description": "Base64 encoded image, returned only if image_exists is true",
          "example": "data:image/png;base64,iVBORw0KGgoAAAANS..."
        },
        "date_creation": {
          "type": "string",
          "format": "date-time"
        },
        "verified": {
          "type": "boolean"
        },
        "Id_User": {
          "type": "string"
        }
      },
      "required": ["content", "visibility", "toxic_score", "date_creation", "Id_User"]
    },
...
    "Comment": {
      "type": "object",
      "properties": {
        "Id_Comment": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "date_upload": {
          "type": "string",
          "format": "date-time"
        },
        "toxic_score": {
          "type": "number",
          "format": "decimal"
        },
        "image_exists": {
          "type": "boolean"
        },
        "image_b64": {
          "type": "string",
          "description": "Base64 encoded image, returned only if image_exists is true",
          "example": "data:image/png;base64,iVBORw0KGgoAAAANS..."
        },
        "verified": {
          "type": "boolean"
        },
        "Id_Reply_to_comment": {
          "type": "string",
          "description": "ID of the comment this comment is replying to"
        },
        "replies": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comment"
          },
          "description": "Replies to this comment"
        },
        "Id_Post": {
          "type": "string"
        },
        "Id_User": {
          "type": "string"
        }
      },
      "required": ["content", "date_upload", "toxic_score", "Id_Post", "Id_User"]
    },
    "NewComment": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "date_upload": {
          "type": "string",
          "format": "date-time"
        },
        "toxic_score": {
          "type": "number",
          "format": "decimal"
        },
        "image_exists": {
          "type": "boolean"
        },
        "image_b64": {
          "type": "string",
          "description": "Base64 encoded image, required only if image_exists is true",
          "example": "data:image/png;base64,iVBORw0KGgoAAAANS..."
        },
        "verified": {
          "type": "boolean"
        },
        "Id_Reply_to_comment": {
          "type": "string",
          "description": "ID of the comment this comment is replying to (optional)"
        },
        "Id_Post": {
          "type": "string"
        },
        "Id_User": {
          "type": "string"
        }
      },
      "required": ["content", "date_upload", "toxic_score", "Id_Post", "Id_User"]
    },
    "Message": {
      "type": "object",
      "properties": {
        "Id_Message": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "date_upload": {
          "type": "string",
          "format": "date-time"
        },
        "Id_User": {
          "type": "string"
        },
        "Id_Conversation": {
          "type": "string",
          "description": "ID of the conversation this message belongs to"
        }
      },
      "required": ["content", "date_upload", "Id_User"]
    },
...
    "APIResult": {
      "type": "object",
      "properties": {
        "Id_api_result": {
          "type": "string"
        },
        "Id_Categories": {
          "type": "integer"
        },
        "flagged": {
          "type": "boolean"
        },
        "Id_Message": {
          "type": "string"
        },
        "Id_Post": {
          "type": "string"
        },
        "Id_Comment": {
          "type": "string"
        },
        "message": {
          "$ref": "#/definitions/Message"
        },
        "post": {
          "$ref": "#/definitions/Post"
        },
        "comment": {
          "$ref": "#/definitions/Comment"
        },
        "Categories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Category"
          }
        }
      },
      "required": ["Id_Categories", "Id_Message", "Id_Post", "Id_Comment"]
    },
    "NewAPIResult": {
      "type": "object",
      "properties": {
        "Id_Categories": {
          "type": "integer"
        },
        "flagged": {
          "type": "boolean"
        },
        "Id_Message": {
          "type": "string"
        },
        "Id_Post": {
          "type": "string"
        },
        "Id_Comment": {
          "type": "string"
        }
      }
    },
    "Category": {
      "type": "object",
      "properties": {
        "Id_Category": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "score": {
          "type": "string"
        },
        "Id_api_result": {
          "type": "string"
        },
        "api_result": {
          "$ref": "#/definitions/APIResult"
        }
      },
      "required": ["name", "Id_api_result"]
    },
    "NewCategory": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "score": {
          "type": "string"
        },
        "Id_api_result": {
          "type": "string"
        }
      }
    },
...
  }
}